requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
httpx==0.28.1
python-dotenv==1.1.1
jsonschema==4.25.1
pytest==8.4.2
